refactor(F2): fold per-variant helpers into their generators
The leftover top-level helper modules are now owned by the variant that uses
them, so the names stop implying they're shared:

- story.py -> folded into generators/baseline.py (run_enhance_chapter,
  run_generate_story_so_far). ~210 lines, under the 300-line module cap.
- world_state.py -> folded into generators/world_state.py (run_init_world_state,
  run_advance_world_state, run_draft_chapter_with_state). ~270 lines.
- story_module.py -> renamed to generators/_dspy_module_impl.py (option b).
  Folding it would collide its 3-field DraftedChapter with
  core.types.DraftedChapter and pull the generator-unused WriteStory/StoryOutline
  reference module into dspy_module.py, so it stays a separate impl module that
  dspy_module.py imports DraftChapter from.

Imports updated (generators/dspy_module.py, test_story_module.py,
tests/test_foundation_reviewer.py) and AGENTS.md File Organization refreshed.
pytest 64 passed; ruff clean; pylint 8.78; vulture/radon clean.
